core/rawdb, ethdb: enable pebble on non-32bit platforms only
