+++ cmd/utils/flags.go
@@ -1017,9 +1017,15 @@ var (
 		AncientFlag,
 		RemoteDBFlag,
 		HttpHeaderFlag,
-		BackingDBFlag,
+		//BackingDBFlag,
 	}
 )
+
+func init() {
+	if rawdb.PebbleEnabled {
+		DatabasePathFlags = append(DatabasePathFlags, BackingDBFlag)
+	}
+}
 
 // MakeDataDir retrieves the currently requested data directory, terminating
 // if none (or the empty string) is specified. If the node is starting a testnet,
